+++ base/shell/explorer/traywnd.cpp
@@ -21,7 +21,6 @@
 
 #include "precomp.h"
 #include <commoncontrols.h>
-#include <traycmd.h>
 
 HRESULT TrayWindowCtxMenuCreator(ITrayWindow * TrayWnd, IN HWND hWndOwner, IContextMenu ** ppCtxMenu);
